move parameters to `parameters.rs`
