+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod ml_kem;
 pub mod utils;
+pub mod parameters;
